requirements.txt
fastapi
uvicorn[standard]
jinja2
aiofiles
python-multipart
python-docx
PyPDF2
openai
python-dotenv
supabase
pydantic